% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahrefs_metrics_data.R
\docType{data}
\name{ahrefs_metrics}
\alias{ahrefs_metrics}
\title{Ahrefs metrics}
\format{An object of class \code{'data.frame'}}
\source{
\href{https://ahrefs.com/api/documentation}{Ahrefs API Documentation}
}
\usage{
data(ahrefs_metrics)
}
\description{
Description of Ahrefs metrics to include in requests with "where" & "having" usage possibility
}
\examples{
data(ahrefs_metrics)
\donttest{View(ahrefs_metrics)}
}
\keyword{datasets}
