% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahrefs_reports_data.R
\docType{data}
\name{ahrefs_reports}
\alias{ahrefs_reports}
\title{ahrefs_reports}
\format{An object of class \code{'data.frame'}}
\source{
\href{https://ahrefs.com/api/documentation}{Ahrefs API Documentation}
}
\usage{
data(ahrefs_reports)
}
\description{
Description of all available Ahrefs reports provided with related function names
}
\examples{
data(ahrefs_reports)
\donttest{View(ahrefs_reports)}
}
\keyword{datasets}
