% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{rah_auth}
\alias{rah_auth}
\title{Authorize your Ahrefs API connection with a API Key (Token)}
\usage{
rah_auth(api_key, verbose = TRUE)
}
\arguments{
\item{api_key}{character string. Valid API key obtained at: https://ahrefs.com/api/profile}

\item{verbose}{logical, defaults to TRUE. Set to FALSE to stop printing status in the console}
}
\value{
invisibly returns API token into environment variable AHREFS_AUTH_TOKEN and prints the status
}
\description{
Authorize your Ahrefs API connection with a API Key (Token)
}
\examples{
\dontrun{
  rah_auth("ABCDEFGHIJKLMNOPQRST")
}
}
